Basic security Spring Boot MYSQL based authorization. Working example ready for implementation. Basic controller only. User management included, DTO added in user management Database created and relations established REST almost finished Error handling in progress

+++ src/main/java/com/sarpio/security/controllers/UserController.java
@@ -2,13 +2,11 @@ package com.sarpio.security.controllers;
 
 import com.sarpio.security.controllers.dto.UserDto;
 import com.sarpio.security.services.UserService;
-import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import javax.transaction.Transactional;
 import javax.validation.Valid;
-import java.util.List;
 
 @RestController
 @RequestMapping("/user")
@@ -17,29 +15,33 @@ public class UserController {
 
     private final UserService userService;
 
-    @Transactional(value = Transactional.TxType.REQUIRED)
-    @PostMapping("/")
-    public UserDto addNewUser(@ApiParam(name = "Add new user") @RequestBody @Valid UserDto dto) {
-        return userService.saveUser(0L, dto);
+//    @Transactional(value = Transactional.TxType.REQUIRED)
+    @PostMapping
+    public ResponseEntity addNewUser(@Valid @RequestBody UserDto dto) {
+        return userService.saveUser(null, dto);
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("{id}")
     public ResponseEntity deleteById(@PathVariable("id") Long id) {
-        return userService.deleteUserById(id);
+        if (userService.getUserById(id)==null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        userService.deleteUserById(id);
+        return new ResponseEntity<>("Deleted User with Id: " + id, HttpStatus.OK);
     }
 
     @GetMapping("/")
-    public List<UserDto> showUsers() {
+    public ResponseEntity<UserDto> showUsers() {
         return userService.showAllUsers();
     }
 
-    @GetMapping("/{id}")
-    public UserDto getUserById(Long id) {
+    @GetMapping("{id}")
+    public ResponseEntity<UserDto> getUserById(@PathVariable("id") Long id) {
         return userService.getUserById(id);
     }
 
-    @PutMapping("/{id}")
-    public UserDto editUser(@Valid @RequestBody UserDto dto, @PathVariable("id") Long id) {
+    @PutMapping("{id}")
+    public ResponseEntity editUser(@Valid @RequestBody UserDto dto, @PathVariable("id") Long id) {
         return userService.saveUser(id, dto);
     }
 }
